Updated project files. Fixed URL scheme.
