+++ app/src/main/res/values/dimens.xml
@@ -1,0 +1,7 @@
+<resources>
+    <!-- Default screen margins, per the Android Design guidelines. -->
+    <dimen name="fab_margin">16dp</dimen>
+    <dimen name="activity_horizontal_margin">16dp</dimen>
+    <dimen name="activity_vertical_margin">64dp</dimen>
+    <dimen name="activity_bottom_margin">16dp</dimen>
+</resources>
