+++ build.gradle
@@ -1,28 +1,23 @@
+// Top-level build file where you can add configuration options common to all sub-projects/modules.
+
 buildscript {
     repositories {
-        mavenCentral()
+        jcenter()
     }
-
     dependencies {
-        classpath 'com.android.tools.build:gradle:0.12.2'
+        classpath 'com.android.tools.build:gradle:2.3.3'
+
+        // NOTE: Do not place your application dependencies here; they belong
+        // in the individual module build.gradle files
     }
 }
 
-apply plugin: 'android'
-
-android {
-    buildToolsVersion "19.1.0"
-    compileSdkVersion 17
+allprojects {
+    repositories {
+        jcenter()
+    }
 }
 
-repositories {
-    mavenCentral()
-}
-
-dependencies {
-    compile fileTree(dir: 'libs', include: '*.jar')
-}
-
-task wrapper(type: Wrapper) {
-    gradleVersion = '1.10'
+task clean(type: Delete) {
+    delete rootProject.buildDir
 }
