--- src/main/res/layout/activity_service_delivery.xml
@@ -1,73 +1,0 @@
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:paddingBottom="@dimen/activity_vertical_margin"
-    android:paddingLeft="@dimen/activity_horizontal_margin"
-    android:paddingRight="@dimen/activity_horizontal_margin"
-    android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context=".ServiceDeliveryActivity" >
-
-    <TextView
-        android:id="@+id/textView1"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_centerHorizontal="true"
-        android:layout_alignParentTop="true"
-        android:text="@string/introduction_service_delivery"
-        android:layout_marginBottom="20dp"
-        android:textAlignment="center" />
-    
-    <TextView
-        android:id="@+id/sp_id_label"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_below="@+id/textView1"
-        android:text="@string/sp_id_label"
-        android:textSize="18sp"
-         />
-
-    <EditText
-        android:id="@+id/sp_id_input"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_alignLeft="@+id/textView1"
-        android:layout_alignRight="@+id/textView1"
-        android:layout_below="@+id/sp_id_label"
-        android:layout_marginBottom="10dp"
-        android:textSize="18sp"
-        android:ems="10" >
-        <requestFocus />
-    </EditText>
-    
-    <TextView
-        android:id="@+id/service_id_label"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_below="@+id/sp_id_input"
-        android:text="@string/service_id_label"
-        android:textSize="18sp"
-        />
-
-    <EditText
-        android:id="@+id/service_id_input"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_alignLeft="@+id/textView1"
-        android:layout_alignRight="@+id/textView1"
-        android:layout_below="@+id/service_id_label"
-        android:layout_marginBottom="10dp"
-        android:textSize="18sp"
-        android:ems="10" >
-    </EditText>
-
-    <Button
-        android:id="@+id/card_button"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_below="@+id/service_id_input"
-        android:layout_centerHorizontal="true"
-        android:layout_marginTop="32dp"
-        android:text="@string/card_button" />
-
-</RelativeLayout>
